Extract shortcut into separate class
